+++ src/Transform.cpp
@@ -53,7 +53,7 @@ Matrix4f getTranslate(float x, float y, float z) {
 Matrix4f getRotateEuler(const AngleValue& x, const AngleValue& y,
                         const AngleValue& z, EulerType type) {
    Matrix4f retVal = Eigen::Matrix4f::Identity();
-   Eigen::Matrix3d block;
+   Eigen::Matrix3f block;
    auto rotX = Eigen::AngleAxisd(x.getRadians(), Eigen::Vector3f::UnitX());
    auto rotY = Eigen::AngleAxisd(y.getRadians(), Eigen::Vector3f::UnitY());
    auto rotZ = Eigen::AngleAxisd(z.getRadians(), Eigen::Vector3f::UnitZ());
